fix(db): DBA audit fixes, migrations, seeds cleanup, admin validation
- Execute migrations 7 (merge statut_validation) and 8 (DBA audit fixes)
- Fix lieu.js model: FLOAT → DECIMAL to match DB schema
- Add statut_validation retro-compatibility in userDTO.js for frontend
- Fix seed-reference-data.sql: remove password_changed_at, set admin validated
- Delete 11 obsolete/dead seed files (PostgreSQL scripts, incompatible SQL, unused JS)
- Rewrite README-DOCKER.md to reflect actual init flow (make setup)
- DBA audit: DECIMAL precision, 14 indexes, bulkCreate, N+1 fixes, CASCADE associations

+++ backend/models/oeuvres/artisanat.js
@@ -33,10 +33,10 @@ module.exports = (sequelize) => {
       type: DataTypes.STRING(255)
     },
     poids: {
-      type: DataTypes.FLOAT
+      type: DataTypes.DECIMAL(8, 3)
     },
     prix: {
-      type: DataTypes.FLOAT
+      type: DataTypes.DECIMAL(10, 2)
     },
     date_creation: {
       type: DataTypes.DATE
